+++ adapters/ji.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for JIProlog 4.1.6.1 or later versions
-%  Last updated on July 12, 2018
+%  Last updated on July 16, 2018
 %
 %  This file is part of Logtalk <https://logtalk.org/>  
 %  Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -390,37 +390,6 @@
 		invoke('java.lang.System', getProperty('java.lang.String'), [Variable], Value),
 		Value \== []
 	).
-
-
-% '$lgt_startup_directory'(-atom)
-%
-% returns the Logtalk startup directory
-
-'$lgt_startup_directory'(Directory) :-
-	(	'$lgt_environment_variable'('LOGTALK_STARTUP_DIRECTORY', Directory0) ->
-		true
-	;	'$lgt_current_directory'(Directory0)
-	),
-	'$lgt_ji_convert_file_path'(Directory0, Directory).
-
-
-% '$lgt_user_directory'(-atom)
-%
-% returns the Logtalk user directory; fails if unknown
-
-'$lgt_user_directory'(Directory) :-
-	'$lgt_environment_variable'('LOGTALKUSER', Directory0),
-	'$lgt_ji_convert_file_path'(Directory0, Directory).
-
-
-
-% '$lgt_home_directory'(-atom)
-%
-% returns the Logtalk home directory; fails if unknown
-
-'$lgt_home_directory'(Directory) :-
-	'$lgt_environment_variable'('LOGTALKHOME', Directory0),
-	'$lgt_ji_convert_file_path'(Directory0, Directory).
 
 
 % '$lgt_decompose_file_name'(+atom, ?atom, ?atom, ?atom)