+++ adapters/swi.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for SWI Prolog 6.6.0 and later versions
-%  Last updated on July 12, 2018
+%  Last updated on July 16, 2018
 %
 %  This file is part of Logtalk <https://logtalk.org/>  
 %  Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -501,35 +501,6 @@
 
 '$lgt_environment_variable'(Variable, Value) :-
 	getenv(Variable, Value).
-
-
-% '$lgt_startup_directory'(-atom)
-%
-% returns the Logtalk startup directory
-
-'$lgt_startup_directory'(Directory) :-
-	(	getenv('LOGTALK_STARTUP_DIRECTORY', Path) ->
-		prolog_to_os_filename(Directory, Path)
-	;	working_directory(Directory, Directory)
-	).
-
-
-% '$lgt_user_directory'(-atom)
-%
-% returns the Logtalk user directory; fails if unknown
-
-'$lgt_user_directory'(Directory) :-
-	getenv('LOGTALKUSER', Path),
-	prolog_to_os_filename(Directory, Path).
-
-
-% '$lgt_home_directory'(-atom)
-%
-% returns the Logtalk home directory; fails if unknown
-
-'$lgt_home_directory'(Directory) :-
-	getenv('LOGTALKHOME', Path),
-	prolog_to_os_filename(Directory, Path).
 
 
 % '$lgt_decompose_file_name'(+atom, ?atom, ?atom, ?atom)