+++ adapters/unsupported/lpawin.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for LPA WinProlog 5.00
-%  Last updated on July 12, 2018
+%  Last updated on July 16, 2018
 %
 %  This file is part of Logtalk <https://logtalk.org/>  
 %  Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -561,36 +561,6 @@ call(F, A1, A2, A3, A4, A5, A6) :-
 '$lgt_environment_variable'(Variable, Value) :-
 	env(Variables),
 	member((Variable,Value), Variables).
-
-
-% '$lgt_startup_directory'(-atom)
-%
-% returns the Logtalk startup directory
-
-'$lgt_startup_directory'(Directory) :-
-	(	env(Variables),
-		member(('LOGTALK_STARTUP_DIRECTORY',Directory), Variables) ->
-		true
-	;	working_directory(Directory)
-	).
-
-
-% '$lgt_user_directory'(-atom)
-%
-% returns the Logtalk user directory; fails if unknown
-
-'$lgt_user_directory'(Directory) :-
-	env(Variables),
-	member(('LOGTALKUSER',Directory), Variables), !.
-
-
-% '$lgt_home_directory'(-atom)
-%
-% returns the Logtalk home directory; fails if unknown
-
-'$lgt_home_directory'(Directory) :-
-	env(Variables),
-	member(('LOGTALKHOME',Directory), Variables), !.
 
 
 % '$lgt_decompose_file_name'(+atom, ?atom, ?atom, ?atom)