+++ adapters/unsupported/tu.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for tuProlog 2.9.0 and later versions
-%  Last updated on July 12, 2018
+%  Last updated on July 16, 2018
 %
 %  This file is part of Logtalk <https://logtalk.org/>  
 %  Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -420,34 +420,6 @@ numbervars(Term, From, Next) :-
 
 '$lgt_environment_variable'(Variable, Value) :-
 	class('java.lang.System') <- getenv(Variable) returns Value.
-
-
-% '$lgt_startup_directory'(-atom)
-%
-% returns the Logtalk startup directory
-
-'$lgt_startup_directory'(Directory) :-
-	class('java.lang.System') <- getenv('LOGTALK_STARTUP_DIRECTORY') returns Directory,
-	(	nonvar(Directory) ->
-		true
-	;	'$lgt_current_directory'(Directory)
-	).
-
-
-% '$lgt_user_directory'(-atom)
-%
-% returns the Logtalk user directory; fails if unknown
-
-'$lgt_user_directory'(Directory) :-
-	class('java.lang.System') <- getenv('LOGTALKUSER') returns Directory.
-
-
-% '$lgt_home_directory'(-atom)
-%
-% returns the Logtalk home directory; fails if unknown
-
-'$lgt_home_directory'(Directory) :-
-	class('java.lang.System') <- getenv('LOGTALKHOME') returns Directory.
 
 
 % '$lgt_decompose_file_name'(+atom, ?atom, ?atom, ?atom)